move MatchMaker in root package
